% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{pensiondatacheck: Efficient and Effective Critical Analysis of Pension Fund Data.}
\usage{
hello(x)
}
\arguments{
\item{x}{description}
}
\value{
The output from \code{\link{print}}
}
\description{
A short description...
}
\examples{
hello("John")
\dontrun{
hello("Steve")

}
}
